# evidencias/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "device_systems API",
    "description": "API REST para la gestión de usuarios de device_systems. Utiliza FastAPI, Pydantic v2 y SQLAlchemy para persistencia relacional, CRUD completo, filtros, validaciones, constraints y manejo controlado de errores. Migraciones Alembic, dispositivos y préstamos.",
    "contact": {
      "name": "Nombre Apellido",
      "email": "ejemplo@device-systems.com"
    },
    "version": "4.0.0"
  },
  "paths": {
...
    "/users/{user_id}/loans": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Préstamos del usuario",
        "description": "Consulta con joins el historial del usuario.",
        "operationId": "user_loans_users__user_id__loans_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Préstamos y equipos relacionados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoanDetailResponse"
                  },
                  "title": "Response User Loans Users  User Id  Loans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/devices/{device_id}/loans": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "Historial del dispositivo",
        "description": "Consulta con joins todos los préstamos históricos de un equipo.",
        "operationId": "device_loans_devices__device_id__loans_get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Préstamos con usuario y dispositivo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoanDetailResponse"
                  },
                  "title": "Response Device Loans Devices  Device Id  Loans Get"
                }
              }
            }
          },
          "404": {
            "description": "Dispositivo inexistente"
          },
          "400": {
            "description": "Serie duplicada o actualización vacía"
          },
          "409": {
            "description": "Dispositivo con préstamos"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/loans/details": {
      "get": {
        "tags": [
          "Loans"
        ],
        "summary": "Consultar préstamos con joins",
        "description": "Relaciona loans, users y devices mediante JOIN; admite los mismos filtros que el listado.",
        "operationId": "list_loans_loans_details_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "device_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LoanStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "user_email",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "email"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Email"
            }
          },
          {
            "name": "device_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Type"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle relacionado de préstamos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoanDetailResponse"
                  },
                  "title": "Response List Loans Loans Details Get"
                }
              }
            }
          },
          "404": {
            "description": "Usuario, dispositivo o préstamo inexistente"
          },
          "409": {
            "description": "Equipo no disponible o préstamo ya devuelto"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/loans": {
      "get": {
        "tags": [
          "Loans"
        ],
        "summary": "Listar y filtrar préstamos",
        "description": "Combina usuario, dispositivo, estado, correo, tipo, búsqueda y rango de fechas UTC inclusivo.",
        "operationId": "list_loans_loans_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "device_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LoanStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "user_email",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "email"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Email"
            }
          },
          {
            "name": "device_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Type"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Préstamos con sus usuarios y dispositivos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoanDetailResponse"
                  },
                  "title": "Response List Loans Loans Get"
                }
              }
            }
          },
          "404": {
            "description": "Usuario, dispositivo o préstamo inexistente"
          },
          "409": {
            "description": "Equipo no disponible o préstamo ya devuelto"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Loans"
        ],
        "summary": "Prestar dispositivo",
        "description": "Valida las referencias y reserva el equipo en la misma transacción del préstamo.",
        "operationId": "create_loan_loans_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Préstamo creado en estado active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuario, dispositivo o préstamo inexistente"
          },
          "409": {
            "description": "Equipo no disponible o préstamo ya devuelto"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/loans/{loan_id}": {
      "get": {
        "tags": [
          "Loans"
        ],
        "summary": "Consultar préstamo",
        "description": "Muestra el préstamo y los datos del usuario y del equipo.",
        "operationId": "get_loan_loans__loan_id__get",
        "parameters": [
          {
            "name": "loan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Loan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del préstamo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuario, dispositivo o préstamo inexistente"
          },
          "409": {
            "description": "Equipo no disponible o préstamo ya devuelto"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/loans/{loan_id}/return": {
      "patch": {
        "tags": [
          "Loans"
        ],
        "summary": "Devolver dispositivo",
        "description": "Marca returned, registra la fecha UTC y libera el equipo en una sola transacción.",
        "operationId": "return_loan_loans__loan_id__return_patch",
        "parameters": [
          {
            "name": "loan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Loan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Préstamo devuelto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuario, dispositivo o préstamo inexistente"
          },
          "409": {
            "description": "Equipo no disponible o préstamo ya devuelto"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DeviceResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "serial_number": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Serial Number"
          },
          "device_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Device Type"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "is_available": {
            "type": "boolean",
            "title": "Is Available",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "serial_number",
          "device_type",
          "id",
          "created_at"
        ],
        "title": "DeviceResponse",
        "example": {
          "brand": "lenovo",
          "device_type": "laptop",
          "name": "Lenovo ThinkPad",
          "serial_number": "LEN-001"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoanCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "User Id"
          },
          "device_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Device Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id",
          "device_id"
        ],
        "title": "LoanCreate",
        "example": {
          "device_id": 1,
          "user_id": 1
        }
      },
      "LoanDetailResponse": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/LoanStatus"
          },
          "return_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Date"
          },
          "user_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "User Id"
          },
          "device_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Device Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "loan_date": {
            "type": "string",
            "format": "date-time",
            "title": "Loan Date"
          },
          "user": {
            "$ref": "#/components/schemas/UserPublic"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceResponse"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "return_date",
          "user_id",
          "device_id",
          "id",
          "loan_date",
          "user",
          "device"
        ],
        "title": "LoanDetailResponse",
        "example": {
          "device_id": 1,
          "user_id": 1
        }
      },
      "LoanResponse": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/LoanStatus"
          },
          "return_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Date"
          },
          "user_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "User Id"
          },
          "device_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Device Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "loan_date": {
            "type": "string",
            "format": "date-time",
            "title": "Loan Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "return_date",
          "user_id",
          "device_id",
          "id",
          "loan_date"
        ],
        "title": "LoanResponse",
        "example": {
          "device_id": 1,
          "user_id": 1
        }
      },
      "LoanStatus": {
        "type": "string",
        "enum": [
          "active",
          "returned",
          "overdue"
        ],
        "title": "LoanStatus"
      },
...
      "UserPublic": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 3,
            "title": "Name",
            "description": "Nombre completo del usuario. Mínimo 3 caracteres.",
            "examples": [
              "Usuario Ejemplo"
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Correo electrónico único del usuario.",
            "examples": [
              "usuario@ejemplo.com"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "description": "Rol del usuario dentro del sistema: admin, support o user.",
            "default": "user"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Indica si el usuario se encuentra activo en el sistema.",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "id"
        ],
        "title": "UserPublic",
        "description": "Response model utilizado por los endpoints de la API pública.",
        "example": {
          "email": "usuario@ejemplo.com",
          "id": 1,
          "is_active": true,
          "name": "Usuario Ejemplo",
          "role": "admin"
        }
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "support",
          "user"
        ],
        "title": "UserRole",
        "description": "Roles permitidos para un usuario del sistema device_systems."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}